% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPMID.R
\name{GetPMID}
\alias{GetPMID}
\title{Extract PMID for a given variable}
\usage{
GetPMID(db, key, value)
}
\arguments{
\item{db}{Character(1). Currently only "cMD" is supported.}

\item{key}{Character(1). Column name where your value belongs to.}

\item{value}{Character(1). Attribute of your interest.}
}
\description{
Extract PMID for a given variable
}
