% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perComp.R
\name{perComp}
\alias{perComp}
\title{Extract the percent completeness of the metadata at the sample level}
\usage{
perComp(db, var)
}
\arguments{
\item{db}{Character(1). Currently only "cMD" is supported.}

\item{var}{Character(1). Name of the variable to check the completeness.}
}
\description{
Extract the percent completeness of the metadata at the sample level
}
